# dtos/personDto.go
package dtos

import "time"

// PersonDto represents data about a person in the people list
type PersonDto struct {
	ID        string    `json:"id"`
	Name      string    `json:"name"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}